Ensure each new session (browser reload for example) gets whole new objects
When a user reloads the page, they get a new session, so they should get all new bokeh and panel objects.
To do this we put the Dashboard class instantiation inside the panel serve so it happens everytime a new session happens.
So that not all sessions share the same Dashboard.